Use auto component delegation for tx components
