+++ crates/relayer-components-extra/src/components/extra/mod.rs
@@ -2,4 +2,3 @@ pub mod build;
 pub mod chain;
 pub mod closures;
 pub mod relay;
-pub mod transaction;
